+++ zellij-server/src/panes/floating_panes/floating_pane_grid.rs
@@ -1,12 +1,8 @@
-use super::pane_resizer::PaneResizer;
 use crate::tab::{MIN_TERMINAL_HEIGHT, MIN_TERMINAL_WIDTH};
 use crate::{panes::PaneId, tab::Pane};
 use std::cmp::Ordering;
-use std::collections::{HashMap, HashSet};
-use zellij_utils::{
-    input::layout::Direction,
-    pane_size::{Dimension, PaneGeom, Size, Viewport},
-};
+use std::collections::HashMap;
+use zellij_utils::pane_size::{Dimension, PaneGeom, Size, Viewport};
 
 use std::cell::RefCell;
 use std::rc::Rc;
@@ -39,11 +35,6 @@ impl<'a> FloatingPaneGrid<'a> {
             display_area,
             viewport,
         }
-    }
-
-    pub fn layout(&mut self, direction: Direction, space: usize) -> Result<(), String> {
-        let mut pane_resizer = PaneResizer::new(self.panes.clone());
-        pane_resizer.layout(direction, space)
     }
     pub fn move_pane_by(&mut self, pane_id: PaneId, x: isize, y: isize) {
         // true => succeeded to move, false => failed to move
@@ -568,16 +559,6 @@ impl<'a> FloatingPaneGrid<'a> {
             None
         }
     }
-    fn increase_pane_height(&mut self, id: &PaneId, percent: f64) {
-        let mut panes = self.panes.borrow_mut();
-        let terminal = panes.get_mut(id).unwrap();
-        terminal.increase_height(percent);
-    }
-    fn increase_pane_width(&mut self, id: &PaneId, percent: f64) {
-        let mut panes = self.panes.borrow_mut();
-        let terminal = panes.get_mut(id).unwrap();
-        terminal.increase_width(percent);
-    }
     fn increase_pane_size_left(&mut self, id: &PaneId, increase_by: usize) {
         let new_pane_geom = {
             let mut panes = self.panes.borrow_mut();
@@ -678,140 +659,6 @@ impl<'a> FloatingPaneGrid<'a> {
             current_geom
         };
         self.set_pane_geom(*id, new_pane_geom);
-    }
-    fn pane_ids_directly_left_of(&self, id: &PaneId) -> Option<Vec<PaneId>> {
-        let panes = self.panes.borrow();
-        let mut ids = vec![];
-        let terminal_to_check = panes.get(id).unwrap();
-        if terminal_to_check.x() == 0 {
-            return None;
-        }
-        // for (&pid, terminal) in self.get_panes() {
-        for (&pid, terminal) in panes.iter() {
-            if terminal.x() + terminal.cols() == terminal_to_check.x() {
-                ids.push(pid);
-            }
-        }
-        if ids.is_empty() {
-            None
-        } else {
-            Some(ids)
-        }
-    }
-    fn pane_ids_directly_right_of(&self, id: &PaneId) -> Option<Vec<PaneId>> {
-        let mut ids = vec![];
-        let panes = self.panes.borrow();
-        let terminal_to_check = panes.get(id).unwrap();
-        // for (&pid, terminal) in self.get_panes() {
-        for (&pid, terminal) in panes.iter() {
-            if terminal.x() == terminal_to_check.x() + terminal_to_check.cols() {
-                ids.push(pid);
-            }
-        }
-        if ids.is_empty() {
-            None
-        } else {
-            Some(ids)
-        }
-    }
-    fn pane_ids_directly_below(&self, id: &PaneId) -> Option<Vec<PaneId>> {
-        let mut ids = vec![];
-        let panes = self.panes.borrow();
-        let terminal_to_check = panes.get(id).unwrap();
-        // for (&pid, terminal) in self.get_panes() {
-        for (&pid, terminal) in panes.iter() {
-            if terminal.y() == terminal_to_check.y() + terminal_to_check.rows() {
-                ids.push(pid);
-            }
-        }
-        if ids.is_empty() {
-            None
-        } else {
-            Some(ids)
-        }
-    }
-    fn pane_ids_directly_above(&self, id: &PaneId) -> Option<Vec<PaneId>> {
-        let mut ids = vec![];
-        let panes = self.panes.borrow();
-        let terminal_to_check = panes.get(id).unwrap();
-        // for (&pid, terminal) in self.get_panes() {
-        for (&pid, terminal) in panes.iter() {
-            if terminal.y() + terminal.rows() == terminal_to_check.y() {
-                ids.push(pid);
-            }
-        }
-        if ids.is_empty() {
-            None
-        } else {
-            Some(ids)
-        }
-    }
-    fn pane_is_between_vertical_borders(
-        &self,
-        id: &PaneId,
-        left_border_x: usize,
-        right_border_x: usize,
-    ) -> bool {
-        let panes = self.panes.borrow();
-        let pane = panes.get(id).unwrap();
-        pane.x() >= left_border_x && pane.x() + pane.cols() <= right_border_x
-    }
-    fn pane_is_between_horizontal_borders(
-        &self,
-        id: &PaneId,
-        top_border_y: usize,
-        bottom_border_y: usize,
-    ) -> bool {
-        let panes = self.panes.borrow();
-        let pane = panes.get(id).unwrap();
-        pane.y() >= top_border_y && pane.y() + pane.rows() <= bottom_border_y
-    }
-    pub fn next_selectable_pane_id(&self, current_pane_id: &PaneId) -> PaneId {
-        let panes = self.panes.borrow();
-        let mut panes: Vec<(&PaneId, &&mut Box<dyn Pane>)> =
-            panes.iter().filter(|(_, p)| p.selectable()).collect();
-        panes.sort_by(|(_a_id, a_pane), (_b_id, b_pane)| {
-            if a_pane.y() == b_pane.y() {
-                a_pane.x().cmp(&b_pane.x())
-            } else {
-                a_pane.y().cmp(&b_pane.y())
-            }
-        });
-        let active_pane_position = panes
-            .iter()
-            .position(|(id, _)| *id == current_pane_id) // TODO: better
-            .unwrap();
-
-        let next_active_pane_id = panes
-            .get(active_pane_position + 1)
-            .or_else(|| panes.get(0))
-            .map(|p| *p.0)
-            .unwrap();
-        next_active_pane_id
-    }
-    pub fn previous_selectable_pane_id(&self, current_pane_id: &PaneId) -> PaneId {
-        let panes = self.panes.borrow();
-        let mut panes: Vec<(&PaneId, &&mut Box<dyn Pane>)> =
-            panes.iter().filter(|(_, p)| p.selectable()).collect();
-        panes.sort_by(|(_a_id, a_pane), (_b_id, b_pane)| {
-            if a_pane.y() == b_pane.y() {
-                a_pane.x().cmp(&b_pane.x())
-            } else {
-                a_pane.y().cmp(&b_pane.y())
-            }
-        });
-        let last_pane = panes.last().unwrap();
-        let active_pane_position = panes
-            .iter()
-            .position(|(id, _)| *id == current_pane_id) // TODO: better
-            .unwrap();
-
-        let previous_active_pane_id = if active_pane_position == 0 {
-            *last_pane.0
-        } else {
-            *panes.get(active_pane_position - 1).unwrap().0
-        };
-        previous_active_pane_id
     }
     pub fn next_selectable_pane_id_to_the_left(&self, current_pane_id: &PaneId) -> Option<PaneId> {
         let panes = self.panes.borrow();
@@ -916,196 +763,6 @@ impl<'a> FloatingPaneGrid<'a> {
             .map(|(_, (pid, _))| pid)
             .copied();
         next_index
-    }
-    fn horizontal_borders(&self, pane_ids: &[PaneId]) -> HashSet<usize> {
-        pane_ids.iter().fold(HashSet::new(), |mut borders, p| {
-            let panes = self.panes.borrow();
-            let pane = panes.get(p).unwrap();
-            borders.insert(pane.y());
-            borders.insert(pane.y() + pane.rows());
-            borders
-        })
-    }
-    fn vertical_borders(&self, pane_ids: &[PaneId]) -> HashSet<usize> {
-        pane_ids.iter().fold(HashSet::new(), |mut borders, p| {
-            let panes = self.panes.borrow();
-            let pane = panes.get(p).unwrap();
-            borders.insert(pane.x());
-            borders.insert(pane.x() + pane.cols());
-            borders
-        })
-    }
-    fn panes_to_the_left_between_aligning_borders(&self, id: PaneId) -> Option<Vec<PaneId>> {
-        let panes = self.panes.borrow();
-        if let Some(pane) = panes.get(&id) {
-            let upper_close_border = pane.y();
-            let lower_close_border = pane.y() + pane.rows();
-
-            if let Some(panes_to_the_left) = self.pane_ids_directly_left_of(&id) {
-                let mut selectable_panes: Vec<_> = panes_to_the_left
-                    .into_iter()
-                    .filter(|pid| panes.get(pid).unwrap().selectable())
-                    .collect();
-                let pane_borders_to_the_left = self.horizontal_borders(&selectable_panes);
-                if pane_borders_to_the_left.contains(&upper_close_border)
-                    && pane_borders_to_the_left.contains(&lower_close_border)
-                {
-                    selectable_panes.retain(|t| {
-                        self.pane_is_between_horizontal_borders(
-                            t,
-                            upper_close_border,
-                            lower_close_border,
-                        )
-                    });
-                    return Some(selectable_panes);
-                }
-            }
-        }
-        None
-    }
-    fn panes_to_the_right_between_aligning_borders(&self, id: PaneId) -> Option<Vec<PaneId>> {
-        let panes = self.panes.borrow();
-        if let Some(pane) = panes.get(&id) {
-            let upper_close_border = pane.y();
-            let lower_close_border = pane.y() + pane.rows();
-
-            if let Some(panes_to_the_right) = self.pane_ids_directly_right_of(&id) {
-                let mut selectable_panes: Vec<_> = panes_to_the_right
-                    .into_iter()
-                    .filter(|pid| panes.get(pid).unwrap().selectable())
-                    .collect();
-                let pane_borders_to_the_right = self.horizontal_borders(&selectable_panes);
-                if pane_borders_to_the_right.contains(&upper_close_border)
-                    && pane_borders_to_the_right.contains(&lower_close_border)
-                {
-                    selectable_panes.retain(|t| {
-                        self.pane_is_between_horizontal_borders(
-                            t,
-                            upper_close_border,
-                            lower_close_border,
-                        )
-                    });
-                    return Some(selectable_panes);
-                }
-            }
-        }
-        None
-    }
-    fn panes_above_between_aligning_borders(&self, id: PaneId) -> Option<Vec<PaneId>> {
-        let panes = self.panes.borrow();
-        if let Some(pane) = panes.get(&id) {
-            let left_close_border = pane.x();
-            let right_close_border = pane.x() + pane.cols();
-
-            if let Some(panes_above) = self.pane_ids_directly_above(&id) {
-                let mut selectable_panes: Vec<_> = panes_above
-                    .into_iter()
-                    .filter(|pid| panes.get(pid).unwrap().selectable())
-                    .collect();
-                let pane_borders_above = self.vertical_borders(&selectable_panes);
-                if pane_borders_above.contains(&left_close_border)
-                    && pane_borders_above.contains(&right_close_border)
-                {
-                    selectable_panes.retain(|t| {
-                        self.pane_is_between_vertical_borders(
-                            t,
-                            left_close_border,
-                            right_close_border,
-                        )
-                    });
-                    return Some(selectable_panes);
-                }
-            }
-        }
-        None
-    }
-    fn panes_below_between_aligning_borders(&self, id: PaneId) -> Option<Vec<PaneId>> {
-        let panes = self.panes.borrow();
-        if let Some(pane) = panes.get(&id) {
-            let left_close_border = pane.x();
-            let right_close_border = pane.x() + pane.cols();
-
-            if let Some(panes_below) = self.pane_ids_directly_below(&id) {
-                let mut selectable_panes: Vec<_> = panes_below
-                    .into_iter()
-                    .filter(|pid| panes[pid].selectable())
-                    .collect();
-                let pane_borders_below = self.vertical_borders(&selectable_panes);
-                if pane_borders_below.contains(&left_close_border)
-                    && pane_borders_below.contains(&right_close_border)
-                {
-                    selectable_panes.retain(|t| {
-                        self.pane_is_between_vertical_borders(
-                            t,
-                            left_close_border,
-                            right_close_border,
-                        )
-                    });
-                    return Some(selectable_panes);
-                }
-            }
-        }
-        None
-    }
-    fn find_panes_to_grow(&self, id: PaneId) -> Option<(Vec<PaneId>, Direction)> {
-        if let Some(panes) = self
-            .panes_to_the_left_between_aligning_borders(id)
-            .or_else(|| self.panes_to_the_right_between_aligning_borders(id))
-        {
-            return Some((panes, Direction::Horizontal));
-        }
-
-        if let Some(panes) = self
-            .panes_above_between_aligning_borders(id)
-            .or_else(|| self.panes_below_between_aligning_borders(id))
-        {
-            return Some((panes, Direction::Vertical));
-        }
-
-        None
-    }
-    fn grow_panes(&mut self, panes: &[PaneId], direction: Direction, (width, height): (f64, f64)) {
-        match direction {
-            Direction::Horizontal => {
-                for pane_id in panes {
-                    self.increase_pane_width(pane_id, width);
-                }
-            }
-            Direction::Vertical => {
-                for pane_id in panes {
-                    self.increase_pane_height(pane_id, height);
-                }
-            }
-        };
-    }
-    pub fn fill_space_over_pane(&mut self, id: PaneId) -> bool {
-        // true => successfully filled space over pane
-        // false => didn't succeed, so didn't do anything
-        let (freed_width, freed_height) = {
-            let panes = self.panes.borrow_mut();
-            let pane_to_close = panes.get(&id).unwrap();
-            let freed_space = pane_to_close.position_and_size();
-            let freed_width = freed_space.cols.as_percent();
-            let freed_height = freed_space.rows.as_percent();
-            (freed_width, freed_height)
-        };
-        if let (Some(freed_width), Some(freed_height)) = (freed_width, freed_height) {
-            if let Some((panes_to_grow, direction)) = self.find_panes_to_grow(id) {
-                self.grow_panes(&panes_to_grow, direction, (freed_width, freed_height));
-                let side_length = match direction {
-                    Direction::Vertical => self.display_area.rows,
-                    Direction::Horizontal => self.display_area.cols,
-                };
-                {
-                    let mut panes = self.panes.borrow_mut();
-                    (*panes).remove(&id);
-                }
-                let mut pane_resizer = PaneResizer::new(self.panes.clone());
-                let _ = pane_resizer.layout(direction, side_length);
-                return true;
-            }
-        }
-        false
     }
     pub fn find_room_for_new_pane(&self) -> Option<PaneGeom> {
         let panes = self.panes.borrow();
