+++ zellij-server/src/panes/tiled_panes/tiled_pane_grid.rs
@@ -1,5 +1,6 @@
 use super::pane_resizer::PaneResizer;
-use crate::tab::{is_inside_viewport, MIN_TERMINAL_HEIGHT, MIN_TERMINAL_WIDTH};
+use super::is_inside_viewport;
+use crate::tab::{MIN_TERMINAL_HEIGHT, MIN_TERMINAL_WIDTH};
 use crate::{panes::PaneId, tab::Pane};
 use std::cmp::Reverse;
 use std::collections::{HashMap, HashSet};
